Refactored views.py and moved all non view functions to utils.py. Moved new page generation code to utilis.py. Deleted old_manage.py as it is now redundant.
